Added UIActionSource. Refactor actions telemetry code. Added additional properties to UISolutionView
