+++ src/NuGet.Clients/NuGet.PackageManagement.UI/Xamls/DetailControl.xaml.cs
@@ -97,7 +97,7 @@ namespace NuGet.PackageManagement.UI
             {
                 var userAction = UserAction.CreateInstallAction(
                     model.Id,
-                    model.SelectedVersion.Version); //diferent from view
+                    model.SelectedVersion.Version, UIActionSource.DetailsPane); //diferent from view
 
                 ExecuteUserAction(userAction, NuGetActionType.Install);
             }
@@ -109,7 +109,7 @@ namespace NuGet.PackageManagement.UI
 
             if (model != null)
             {
-                var userAction = UserAction.CreateUnInstallAction(model.Id);
+                var userAction = UserAction.CreateUnInstallAction(model.Id, UIActionSource.DetailsPane);
                 ExecuteUserAction(userAction, NuGetActionType.Uninstall);
             }
         }
@@ -122,7 +122,7 @@ namespace NuGet.PackageManagement.UI
             {
                 var userAction = UserAction.CreateInstallAction(
                     model.Id,
-                    model.SelectedVersion.Version);
+                    model.SelectedVersion.Version, UIActionSource.DetailsPane);
 
                 ExecuteUserAction(userAction, NuGetActionType.Install);
             }
@@ -134,7 +134,7 @@ namespace NuGet.PackageManagement.UI
 
             if (model != null)
             {
-                var userAction = UserAction.CreateUnInstallAction(model.Id);
+                var userAction = UserAction.CreateUnInstallAction(model.Id, UIActionSource.DetailsPane);
                 ExecuteUserAction(userAction, NuGetActionType.Uninstall);
             }
         }
